+++ FusionEngine/src/Platform/Vulkan/VulkanSwapChain.cpp
@@ -174,7 +174,7 @@ namespace FusionEngine
         CreateSwapChain();
         CreateFrameBuffers();
 
-        m_RenderApi->m_ResourceManager->SetFrameCount(m_RenderApi->m_Frames.size());
+        m_RenderApi->m_ResourceManager->SetFrameCount(static_cast<uint32_t>(m_RenderApi->m_Frames.size()));
         for (auto& frame : m_RenderApi->m_Frames)
         {
             frame.CommandBuffer = m_RenderApi->CreateCommandBuffer();
